requirements.txt
streamlit
torch
numpy
Pillow 